+++ src/gdb/remote.h
@@ -59,9 +59,9 @@ extern void remote_cisco_objfile_relocate(bfd_signed_vma text_off,
 extern void async_remote_interrupt_twice(void *arg);
 
 extern int remote_write_bytes(CORE_ADDR memaddr, const gdb_byte *myaddr,
-                              int len);
+                              size_t len);
 
-extern int remote_read_bytes(CORE_ADDR memaddr, char *myaddr, int len);
+extern int remote_read_bytes(CORE_ADDR memaddr, char *myaddr, size_t len);
 
 extern void (*deprecated_target_resume_hook)(void)
   ATTRIBUTE_DEPRECATED;
